+++ src/main/java/nik/trade/tradeapp2/repository/OrderRepository.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 public interface OrderRepository extends MongoRepository<Order,String> {
-    List<Order> findByPriseSale(int priseSale);
+    List<Order> findByPriseSale(double priseSale);
     List<Order> findByDate (LocalDate date);
     List<Order> findByCustomer (Customer customer);
 }
